Added file reorienter and associated tools
file_reorient will reorient a file from the specified orientation to any new orientation that is required. It uses a couple of small functions from file_tools as well. Feel free to add functionality that adds specific files as possible arguments instead of changing the code directly if one uses it with another pipeline.
Most of the functions originate from DTC library